+++ dcm_import_module/__init__.py
@@ -45,7 +45,9 @@ def app_factory(
 
     # create Orchestrator and OrchestratedView-classes
     orchestrator = ScalableOrchestrator(
-        queue=queue or config.queue, registry=registry or config.registry
+        queue=queue or config.queue,
+        registry=registry or config.registry,
+        nprocesses=config.ORCHESTRATION_PROCESSES,
     )
     view_external = ExternalImportView(
         config=config,