+++ dcm_import_module/plugins/oai_pmh.py
@@ -1,5 +1,8 @@
 """OAI-protocol based IE-import plugin."""
 
+from typing import Optional, Mapping
+from random import sample
+
 import requests
 from dcm_common.logger import LoggingContext as Context, Logger
 from dcm_common.plugins import PythonDependency, Signature, Argument, JSONType
@@ -33,6 +36,7 @@ class OAIPMHPlugin(IEImportPlugin):
     ]
 
     _SIGNATURE = Signature(
+        test=IEImportPlugin.signature.properties["test"],
         transfer_url_info=Argument(
             type_=JSONType.OBJECT,
             required=True,
@@ -122,6 +126,34 @@ class OAIPMHPlugin(IEImportPlugin):
         ):
             return False, f"Bad url-scheme, supported are {qjoin(schemes)}"
         return True, ""
+
+    def _get_collector(self, transfer_url_info: Mapping) -> PayloadCollector:
+        if "xml_path" in transfer_url_info:
+            return PayloadCollector(
+                TransferUrlFilters.filter_by_regex_in_xml_path(
+                    transfer_url_info["regex"],
+                    transfer_url_info["xml_path"],
+                ),
+                self._timeout,
+            )
+        return PayloadCollector(
+            TransferUrlFilters.filter_by_regex(transfer_url_info["regex"]),
+            self._timeout,
+        )
+
+    def _list_identifiers(
+        self,
+        interface: RepositoryInterface,
+        request_args: Mapping,
+        /,
+        **kwargs,
+    ) -> tuple[Logger, Optional[list[str]]]:
+        return self._retry(
+            interface.list_identifiers_exhaustive,
+            kwargs=request_args,
+            description="collecting identifiers",
+            exceptions=requests.exceptions.ReadTimeout,
+        )
 
     def _get_records(
         self,
@@ -227,21 +259,7 @@ class OAIPMHPlugin(IEImportPlugin):
         # initialize from request-args
         interface = RepositoryInterface(kwargs["base_url"], self._timeout)
         interface.preserve_log = True
-        if "xml_path" in kwargs["transfer_url_info"]:
-            collector = PayloadCollector(
-                TransferUrlFilters.filter_by_regex_in_xml_path(
-                    kwargs["transfer_url_info"]["regex"],
-                    kwargs["transfer_url_info"]["xml_path"],
-                ),
-                self._timeout,
-            )
-        else:
-            collector = PayloadCollector(
-                TransferUrlFilters.filter_by_regex(
-                    kwargs["transfer_url_info"]["regex"]
-                ),
-                self._timeout,
-            )
+        collector = self._get_collector(kwargs["transfer_url_info"])
 
         # fetch identifiers
         if "identifiers" in kwargs:
@@ -265,27 +283,48 @@ class OAIPMHPlugin(IEImportPlugin):
             if "set_spec" in kwargs:
                 request_args["_set_spec"] = kwargs["set_spec"]
 
-            log, identifiers = self._retry(
-                interface.list_identifiers_exhaustive,
-                kwargs=request_args,
-                description="collecting identifiers",
-                exceptions=requests.exceptions.ReadTimeout,
+            log, identifiers = self._list_identifiers(
+                interface, request_args, **kwargs
             )
+            context.result.log.merge(log)
             if Context.ERROR in log:
                 context.set_progress("collecting identifiers failed")
-                context.result.log.merge(log)
                 context.result.log.log(
                     Context.ERROR, body="Collecting identifiers failed."
                 )
                 context.push()
                 return context.result
 
+            # test-import: limit number of identifiers
+            if (
+                kwargs.get("test", False)
+                and identifiers is not None
+                and len(identifiers) > self._test_volume >= 0
+            ):
+                context.result.log.log(
+                    Context.INFO,
+                    body=(
+                        "Limiting number of identifiers from "
+                        + f"{len(identifiers)} down to {self._test_volume} "
+                        + f"via '{self._test_strategy}'-strategy."
+                    ),
+                )
+                context.push()
+                match self._test_strategy:
+                    case "first":
+                        identifiers = identifiers[: self._test_volume]
+                    case "random":
+                        identifiers = sample(identifiers, k=self._test_volume)
+                    case _:
+                        raise ValueError(
+                            f"Unknown test-strategy '{self._test_strategy}'."
+                        )
+
         # process identifiers
         if len(identifiers) == 0:
             context.result.log.log(
                 Context.WARNING, body="List of identifiers is empty."
             )
-            context.result.log.merge(interface.log)
             context.push()
         else:
             # download records
@@ -306,3 +345,111 @@ class OAIPMHPlugin(IEImportPlugin):
             context.set_progress("failure")
         context.push()
         return context.result
+
+
+class OAIPMHPlugin2(OAIPMHPlugin):
+    """
+    IE-import plugin based on the OAI-protocol [1].
+
+    It inherits from the `OAIPMHPlugin` and adds support for multiple sets,
+    and multiple transferUrlFilters based on xpath.
+
+    [1] https://www.openarchives.org/OAI/openarchivesprotocol.html
+    """
+
+    _NAME = "oai_pmh_v2"
+    _DESCRIPTION = (
+        "IE import based on the Open Archives Initiative protocol for "
+        + "metadata harvesting (OAI-PMH). "
+        + "Supports multiple sets and transferUrlFilters."
+    )
+    _SIGNATURE = Signature(
+        test=IEImportPlugin.signature.properties["test"],
+        base_url=OAIPMHPlugin.signature.properties["base_url"],
+        metadata_prefix=OAIPMHPlugin.signature.properties["metadata_prefix"],
+        identifiers=OAIPMHPlugin.signature.properties["identifiers"],
+        from_=OAIPMHPlugin.signature.properties["from_"],
+        until=OAIPMHPlugin.signature.properties["until"],
+        transfer_url_info=Argument(
+            type_=JSONType.ARRAY,
+            required=True,
+            item_type=Argument(
+                type_=JSONType.OBJECT,
+                required=True,
+                description=(
+                    "information that allows to extract payload urls from "
+                    + "source metadata"
+                ),
+                properties={
+                    "path": Argument(
+                        JSONType.STRING,
+                        required=False,
+                        item_type=JSONType.STRING,
+                        description=(
+                            "xpath query; "
+                            + "if omitted, the regex is applied to the entire "
+                            + "source metadata XML"
+                        ),
+                        example=(
+                            "./GetRecord/record/metadata/"
+                            + "oai_dc:dc/dc:identifier"
+                        ),
+                        default=None,
+                    ),
+                    "regex": Argument(
+                        JSONType.STRING,
+                        required=True,
+                        description=(
+                            "regex used to match content of 'path' or "
+                            + "full source metadata; based on 're.findall'"
+                        ),
+                        example=(
+                            r"(https://repositorium\.uni-muenster\.de/transfer/.*)"
+                        ),
+                    ),
+                },
+            ),
+        ),
+        set_spec=Argument(
+            type_=JSONType.ARRAY,
+            required=False,
+            default=None,
+            item_type=JSONType.STRING,
+            description=(
+                "selective harvesting (list of set memberships; "
+                + "identifiers combined via a logical OR-operation)"
+            ),
+            example=["doc-type:article"],
+        ),
+    )
+
+    def _get_collector(self, transfer_url_info: Mapping) -> PayloadCollector:
+        return PayloadCollector(
+            transfer_url_filters=list(
+                map(
+                    lambda info: (
+                        TransferUrlFilters.filter_by_regex_with_xpath_query(
+                            **info
+                        )
+                        if "path" in info
+                        else TransferUrlFilters.filter_by_regex(**info)
+                    ),
+                    transfer_url_info,
+                )
+            ),
+            timeout=self._timeout
+        )
+
+    def _list_identifiers(
+        self,
+        interface: RepositoryInterface,
+        request_args: Mapping,
+        /,
+        **kwargs,
+    ) -> tuple[Logger, Optional[list[str]]]:
+        return self._retry(
+            interface.list_identifiers_exhaustive_multiple_sets,
+            kwargs=request_args,
+            description="collecting identifiers",
+            exceptions=requests.exceptions.ReadTimeout,
+        )